feat: useVariableStore separated into two files
