+++ src/components/react-components/model-editor/ChangeUpdateFunctionInput/ChangeUpdateFunctionInput.tsx
@@ -1,5 +1,5 @@
 import useUpdateFunctionsStore from '../../../../stores/LiveModel/useUpdateFunctionsStore';
-import useVariablesStore from '../../../../stores/LiveModel/useVariablesStore';
+import useVariablesStore from '../../../../stores/LiveModel/VariableStore/useVariablesStore';
 import InvisibleInputReact from '../../lit-wrappers/InvisibleInputReact';
 import type { ChangeUpdateFunctionInputProps } from './ChangeUpdateFunctionInputProps';
